website/requirements.txt
Flask
flask-sqlalchemy
python-dotenv
python-json-logger
pandas
flask-login
bcrypt
requests
gunicorn